Order Activity sorta works...

+++ src/uc/tjt/estadium/Consumable.java
@@ -3,11 +3,11 @@ package uc.tjt.estadium;
 import android.graphics.Bitmap;
 
 public class Consumable{
-	int ID;
+	long ID;
 	String name;
 	float price;
 	Bitmap mIcon;
-	Consumable(int id,String n,float p){
+	Consumable(long id,String n,float p){
 		name=n;price=p;ID=id;
 	}
 }
